+++ packages/cloud_firestore/cloud_firestore_web/lib/src/collection_reference_web.dart
@@ -3,12 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:firebase/firestore.dart' as web;
 import 'package:meta/meta.dart';
 
-import 'package:cloud_firestore_web/src/document_reference_web.dart';
-import 'package:cloud_firestore_web/src/query_web.dart';
-
+import 'interop/firestore.dart' as firestore_interop;
 import 'document_reference_web.dart';
 import 'query_web.dart';
 
@@ -16,12 +13,12 @@ import 'query_web.dart';
 class CollectionReferenceWeb extends QueryWeb
     implements CollectionReferencePlatform {
   /// instance of Firestore from the web plugin
-  final web.Firestore _webFirestore;
+  final firestore_interop.Firestore _webFirestore;
 
   final FirebaseFirestorePlatform _firestorePlatform;
 
   /// instance of DocumentReference from the web plugin
-  final web.CollectionReference _delegate;
+  final firestore_interop.CollectionReference _delegate;
 
   // disabling lint as it's only visible for testing
   @visibleForTesting
@@ -39,7 +36,7 @@ class CollectionReferenceWeb extends QueryWeb
 
   @override
   DocumentReferencePlatform doc([String path]) {
-    web.DocumentReference documentReference = _delegate.doc(path);
+    firestore_interop.DocumentReference documentReference = _delegate.doc(path);
     return DocumentReferenceWeb(
         _firestorePlatform, _webFirestore, documentReference.path);
   }
@@ -49,7 +46,7 @@ class CollectionReferenceWeb extends QueryWeb
 
   @override
   DocumentReferencePlatform get parent {
-    web.DocumentReference documentReference = _delegate.parent;
+    firestore_interop.DocumentReference documentReference = _delegate.parent;
 
     if (documentReference == null) {
       return null;