+++ packages/cloud_firestore/cloud_firestore_web/lib/src/utils/web_utils.dart
@@ -1,11 +1,11 @@
-// Copyright 2017, the Chromium project authors.  Please see the AUTHORS file
+// Copyright 2020, the Chromium project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:firebase/firestore.dart' as web;
 
-import 'package:cloud_firestore_web/src/utils/codec_utility.dart';
+import '../utils/codec_utility.dart';
+import '../interop/firestore.dart' as firestore_interop;
 
 const _kChangeTypeAdded = "added";
 const _kChangeTypeModified = "modified";
@@ -13,7 +13,8 @@ const _kChangeTypeRemoved = "removed";
 
 /// Converts a [web.QuerySnapshot] to a [QuerySnapshotPlatform].
 QuerySnapshotPlatform convertWebQuerySnapshot(
-    FirebaseFirestorePlatform firestore, web.QuerySnapshot webQuerySnapshot) {
+    FirebaseFirestorePlatform firestore,
+    firestore_interop.QuerySnapshot webQuerySnapshot) {
   return QuerySnapshotPlatform(
     webQuerySnapshot.docs
         .map((webDocumentSnapshot) =>
@@ -30,7 +31,8 @@ QuerySnapshotPlatform convertWebQuerySnapshot(
 
 /// Converts a [web.DocumentSnapshot] to a [DocumentSnapshotPlatform].
 DocumentSnapshotPlatform convertWebDocumentSnapshot(
-    FirebaseFirestorePlatform firestore, web.DocumentSnapshot webSnapshot) {
+    FirebaseFirestorePlatform firestore,
+    firestore_interop.DocumentSnapshot webSnapshot) {
   return DocumentSnapshotPlatform(
     firestore,
     webSnapshot.ref.path,
@@ -46,7 +48,8 @@ DocumentSnapshotPlatform convertWebDocumentSnapshot(
 
 /// Converts a [web.DocumentChange] to a [DocumentChangePlatform].
 DocumentChangePlatform convertWebDocumentChange(
-    FirebaseFirestorePlatform firestore, web.DocumentChange webDocumentChange) {
+    FirebaseFirestorePlatform firestore,
+    firestore_interop.DocumentChange webDocumentChange) {
   return (DocumentChangePlatform(
       convertWebDocumentChangeType(webDocumentChange.type),
       webDocumentChange.oldIndex,
@@ -70,7 +73,54 @@ DocumentChangeType convertWebDocumentChangeType(String changeType) {
 
 /// Converts a [web.SnapshotMetadata] to a [SnapshotMetadataPlatform].
 SnapshotMetadataPlatform convertWebSnapshotMetadata(
-    web.SnapshotMetadata webSnapshotMetadata) {
+    firestore_interop.SnapshotMetadata webSnapshotMetadata) {
   return SnapshotMetadataPlatform(
       webSnapshotMetadata.hasPendingWrites, webSnapshotMetadata.fromCache);
+}
+
+/// Converts a [GetOptions] to a [web.GetOptions].
+firestore_interop.GetOptions convertGetOptions(GetOptions options) {
+  if (options == null) return null;
+
+  var source;
+  if (options.source != null) {
+    switch (options.source) {
+      case Source.serverAndCache:
+        source = 'default';
+        break;
+      case Source.cache:
+        source = 'cache';
+        break;
+      case Source.server:
+        source = 'server';
+        break;
+      default:
+        source = 'default';
+        break;
+    }
+  }
+
+  return firestore_interop.GetOptions(source: source);
+}
+
+/// Converts a [SetOptions] to a [web.SetOptions].
+firestore_interop.SetOptions convertSetOptions(SetOptions options) {
+  if (options == null) return null;
+
+  var parsedOptions;
+  if (options.merge != null) {
+    parsedOptions = firestore_interop.SetOptions(merge: options.merge);
+  } else if (options.mergeFields != null) {
+    parsedOptions = firestore_interop.SetOptions(
+        mergeFields: options.mergeFields
+            .map((e) => e.components.toList().join('.'))
+            .toList());
+  }
+
+  return parsedOptions;
+}
+
+/// Converts a [FieldPath] to a [web.FieldPath].
+firestore_interop.FieldPath convertFieldPath(FieldPath fieldPath) {
+  return firestore_interop.FieldPath(fieldPath.components.toList().join('.'));
 }