+++ packages/firebase_auth/firebase_auth_web/lib/src/utils/web_utils.dart
@@ -3,43 +3,52 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart';
-import 'package:firebase/firebase.dart' as firebase;
+import '../interop/auth.dart' as auth_interop;
+import 'package:firebase_core_web/firebase_core_web_interop.dart'
+    as core_interop;
 
 /// Given a web error, a [FirebaseAuthException] is returned.
 ///
-/// The firebase-dart wrapper exposes a [firebase.FirebaseError], allowing us to
+/// The firebase-dart wrapper exposes a [core_interop.FirebaseError], allowing us to
 /// use the code and message and convert it into an expected [FirebaseAuthException].
-///
-/// TODO: The firebase-dart wrapper does not support email or credential properties.
-FirebaseAuthException throwFirebaseAuthException(Object exception) {
-  if (exception is! firebase.FirebaseError) {
+FirebaseAuthException getFirebaseAuthException(Object exception) {
+  if (exception is! core_interop.FirebaseError) {
     return FirebaseAuthException(
         code: 'unknown', message: 'An unknown error occurred: ${exception}');
   }
 
-  firebase.FirebaseError firebaseError = exception as firebase.FirebaseError;
+  auth_interop.AuthError firebaseError = exception as auth_interop.AuthError;
 
   String code = firebaseError.code.replaceFirst('auth/', '');
   String message =
       firebaseError.message.replaceFirst('(${firebaseError.code})', '');
-  return FirebaseAuthException(code: code, message: message);
+
+  return FirebaseAuthException(
+    code: code,
+    message: message,
+    email: firebaseError.email,
+    credential: convertWebAuthCredential(firebaseError.credential),
+    phoneNumber: firebaseError.phoneNumber,
+    tenantId: firebaseError.tenantId,
+  );
 }
 
-/// Converts a [firebase.ActionCodeInfo] into a [ActionCodeInfo].
+/// Converts a [auth_interop.ActionCodeInfo] into a [ActionCodeInfo].
 ActionCodeInfo convertWebActionCodeInfo(
-    firebase.ActionCodeInfo webActionCodeInfo) {
+    auth_interop.ActionCodeInfo webActionCodeInfo) {
   if (webActionCodeInfo == null) {
     return null;
   }
-  // TODO: firebase-dart missing operation, previousEmail - defaulting to 'unknown'.
+
   return ActionCodeInfo(operation: 0, data: <String, dynamic>{
     'email': webActionCodeInfo.data.email,
+    'previousEmail': webActionCodeInfo.data.previousEmail,
   });
 }
 
-/// Converts a [firebase.AdditionalUserInfo] into a [AdditionalUserInfo].
+/// Converts a [auth_interop.AdditionalUserInfo] into a [AdditionalUserInfo].
 AdditionalUserInfo convertWebAdditionalUserInfo(
-    firebase.AdditionalUserInfo webAdditionalUserInfo) {
+    auth_interop.AdditionalUserInfo webAdditionalUserInfo) {
   if (webAdditionalUserInfo == null) {
     return null;
   }
@@ -52,8 +61,9 @@ AdditionalUserInfo convertWebAdditionalUserInfo(
   );
 }
 
-/// Converts a [firebase.IdTokenResult] into a [IdTokenResult].
-IdTokenResult convertWebIdTokenResult(firebase.IdTokenResult webIdTokenResult) {
+/// Converts a [auth_interop.IdTokenResult] into a [IdTokenResult].
+IdTokenResult convertWebIdTokenResult(
+    auth_interop.IdTokenResult webIdTokenResult) {
   return IdTokenResult(<String, dynamic>{
     'claims': webIdTokenResult.claims,
     'expirationTimestamp':
@@ -66,34 +76,33 @@ IdTokenResult convertWebIdTokenResult(firebase.IdTokenResult webIdTokenResult) {
   });
 }
 
-/// Converts a [ActionCodeSettings] into a [firebase.ActionCodeSettings].
-firebase.ActionCodeSettings convertPlatformActionCodeSettings(
+/// Converts a [ActionCodeSettings] into a [auth_interop.ActionCodeSettings].
+auth_interop.ActionCodeSettings convertPlatformActionCodeSettings(
     ActionCodeSettings actionCodeSettings) {
   if (actionCodeSettings == null) {
     return null;
   }
 
   Map<String, dynamic> actionCodeSettingsMap = actionCodeSettings.asMap();
-  firebase.AndroidSettings android;
-  firebase.IosSettings iOS;
+
+  auth_interop.ActionCodeSettings webActionCodeSettings =
+      auth_interop.ActionCodeSettings(
+          url: actionCodeSettings.url,
+          handleCodeInApp: actionCodeSettings.handleCodeInApp);
 
   if (actionCodeSettingsMap['android'] != null) {
-    android = firebase.AndroidSettings(
+    webActionCodeSettings.android = auth_interop.AndroidSettings(
         packageName: actionCodeSettingsMap['android']['packageName'],
         minimumVersion: actionCodeSettingsMap['android']['minimumVersion'],
         installApp: actionCodeSettingsMap['android']['installApp']);
   }
 
   if (actionCodeSettingsMap['iOS'] != null) {
-    iOS = firebase.IosSettings(
+    webActionCodeSettings.iOS = auth_interop.IosSettings(
         bundleId: actionCodeSettingsMap['iOS']['bundleId']);
   }
 
-  return firebase.ActionCodeSettings(
-      url: actionCodeSettings.url,
-      handleCodeInApp: actionCodeSettings.handleCodeInApp,
-      android: android,
-      iOS: iOS);
+  return webActionCodeSettings;
 }
 
 /// Converts a [Persistence] enum into a web string persistence value.
@@ -109,15 +118,16 @@ String convertPlatformPersistence(Persistence persistence) {
   }
 }
 
-/// Converts a [AuthProvider] into a [firebase.AuthProvider].
-firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
+/// Converts a [AuthProvider] into a [auth_interop.AuthProvider].
+auth_interop.AuthProvider convertPlatformAuthProvider(
+    AuthProvider authProvider) {
   if (authProvider is EmailAuthProvider) {
-    return firebase.EmailAuthProvider();
+    return auth_interop.EmailAuthProvider();
   }
 
   if (authProvider is FacebookAuthProvider) {
-    firebase.FacebookAuthProvider facebookAuthProvider =
-        firebase.FacebookAuthProvider();
+    auth_interop.FacebookAuthProvider facebookAuthProvider =
+        auth_interop.FacebookAuthProvider();
 
     authProvider.scopes
         .forEach((String scope) => facebookAuthProvider.addScope(scope));
@@ -127,8 +137,8 @@ firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
   }
 
   if (authProvider is GithubAuthProvider) {
-    firebase.GithubAuthProvider githubAuthProvider =
-        firebase.GithubAuthProvider();
+    auth_interop.GithubAuthProvider githubAuthProvider =
+        auth_interop.GithubAuthProvider();
 
     authProvider.scopes
         .forEach((String scope) => githubAuthProvider.addScope(scope));
@@ -138,8 +148,8 @@ firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
   }
 
   if (authProvider is GoogleAuthProvider) {
-    firebase.GoogleAuthProvider googleAuthProvider =
-        firebase.GoogleAuthProvider();
+    auth_interop.GoogleAuthProvider googleAuthProvider =
+        auth_interop.GoogleAuthProvider();
 
     authProvider.scopes
         .forEach((String scope) => googleAuthProvider.addScope(scope));
@@ -149,8 +159,8 @@ firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
   }
 
   if (authProvider is TwitterAuthProvider) {
-    firebase.TwitterAuthProvider twitterAuthProvider =
-        firebase.TwitterAuthProvider();
+    auth_interop.TwitterAuthProvider twitterAuthProvider =
+        auth_interop.TwitterAuthProvider();
 
     twitterAuthProvider.setCustomParameters(
         Map<String, dynamic>.from(authProvider.parameters));
@@ -158,12 +168,12 @@ firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
   }
 
   if (authProvider is PhoneAuthProvider) {
-    return firebase.PhoneAuthProvider();
+    return auth_interop.PhoneAuthProvider();
   }
 
   if (authProvider is OAuthProvider) {
-    firebase.OAuthProvider oAuthProvider =
-        firebase.OAuthProvider(authProvider.providerId);
+    auth_interop.OAuthProvider oAuthProvider =
+        auth_interop.OAuthProvider(authProvider.providerId);
 
     authProvider.scopes
         .forEach((String scope) => oAuthProvider.addScope(scope));
@@ -175,9 +185,22 @@ firebase.AuthProvider convertPlatformAuthProvider(AuthProvider authProvider) {
   return null;
 }
 
-/// Converts a [firebase.OAuthCredential] into a [AuthCredential].
+/// Converts a [auth_interop.AuthCredential] into a [AuthCredential].
+AuthCredential convertWebAuthCredential(
+    auth_interop.AuthCredential authCredential) {
+  if (authCredential == null) {
+    return null;
+  }
+
+  return AuthCredential(
+    providerId: authCredential.providerId,
+    signInMethod: authCredential.signInMethod,
+  );
+}
+
+/// Converts a [auth_interop.OAuthCredential] into a [AuthCredential].
 AuthCredential convertWebOAuthCredential(
-    firebase.OAuthCredential oAuthCredential) {
+    auth_interop.OAuthCredential oAuthCredential) {
   if (oAuthCredential == null) {
     return null;
   }
@@ -189,47 +212,47 @@ AuthCredential convertWebOAuthCredential(
 }
 
 /// Converts a [AuthCredential] into a [firebase.OAuthCredential].
-firebase.OAuthCredential convertPlatformCredential(AuthCredential credential) {
+auth_interop.OAuthCredential convertPlatformCredential(
+    AuthCredential credential) {
   if (credential is EmailAuthCredential) {
     if (credential.emailLink != null) {
-      // TODO: not supported by firebase-dart
-      throw UnimplementedError(
-          "EmailAuthProvider.credentialWithLink() is not supported on web");
+      return auth_interop.EmailAuthProvider.credentialWithLink(
+          credential.email, credential.emailLink);
     }
-    return firebase.EmailAuthProvider.credential(
+    return auth_interop.EmailAuthProvider.credential(
         credential.email, credential.password);
   }
 
   if (credential is FacebookAuthCredential) {
-    return firebase.FacebookAuthProvider.credential(credential.accessToken);
+    return auth_interop.FacebookAuthProvider.credential(credential.accessToken);
   }
 
   if (credential is GithubAuthCredential) {
-    return firebase.GithubAuthProvider.credential(credential.accessToken);
+    return auth_interop.GithubAuthProvider.credential(credential.accessToken);
   }
 
   if (credential is GoogleAuthCredential) {
-    return firebase.GoogleAuthProvider.credential(
+    return auth_interop.GoogleAuthProvider.credential(
         credential.idToken, credential.accessToken);
   }
 
   if (credential is OAuthCredential) {
-    return firebase.OAuthProvider(credential.providerId)
+    return auth_interop.OAuthProvider(credential.providerId)
         .credential(credential.idToken, credential.accessToken);
   }
 
   if (credential is TwitterAuthCredential) {
-    return firebase.TwitterAuthProvider.credential(
+    return auth_interop.TwitterAuthProvider.credential(
         credential.accessToken, credential.secret);
   }
 
   if (credential is PhoneAuthCredential) {
-    return firebase.PhoneAuthProvider.credential(
+    return auth_interop.PhoneAuthProvider.credential(
         credential.verificationId, credential.smsCode);
   }
 
   if (credential is OAuthCredential) {
-    return firebase.OAuthProvider(credential.providerId)
+    return auth_interop.OAuthProvider(credential.providerId)
         .credential(credential.idToken, credential.accessToken);
   }