+++ packages/firebase_messaging/firebase_messaging/example/lib/main.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
@@ -137,7 +138,9 @@ class _Application extends State<Application> {
                 channel.id,
                 channel.name,
                 channel.description,
-                icon: android?.smallIcon,
+                // TODO add a proper drawable resource to android, for now using
+                //      one that already exists in example app.
+                icon: 'launch_background',
               ),
             ));
       }
@@ -170,11 +173,70 @@ class _Application extends State<Application> {
     }
   }
 
+  void onActionSelected(String value) async {
+    switch (value) {
+      case 'subscribe':
+        {
+          print(
+              'FlutterFire Messaging Example: Subscribing to topic "fcm_test".');
+          await FirebaseMessaging.instance.subscribeToTopic('fcm_test');
+          print(
+              'FlutterFire Messaging Example: Subscribing to topic "fcm_test" successful.');
+        }
+        break;
+      case 'unsubscribe':
+        {
+          print(
+              'FlutterFire Messaging Example: Unsubscribing from topic "fcm_test".');
+          await FirebaseMessaging.instance.unsubscribeFromTopic('fcm_test');
+          print(
+              'FlutterFire Messaging Example: Unsubscribing from topic "fcm_test" successful.');
+        }
+        break;
+      case 'get_apns_token':
+        {
+          if (defaultTargetPlatform == TargetPlatform.iOS ||
+              defaultTargetPlatform == TargetPlatform.macOS) {
+            print('FlutterFire Messaging Example: Getting APNs token...');
+            String token = await FirebaseMessaging.instance.getAPNSToken();
+            print('FlutterFire Messaging Example: Got APNs token: $token');
+          } else {
+            print(
+                'FlutterFire Messaging Example: Getting an APNs token is only supported on iOS and macOS platforms.');
+          }
+        }
+        break;
+      default:
+        break;
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text("Cloud Messaging"),
+        actions: <Widget>[
+          PopupMenuButton(
+            onSelected: onActionSelected,
+            itemBuilder: (BuildContext context) {
+              return [
+                PopupMenuItem(
+                  value: "subscribe",
+                  child: Text("Subscribe to topic"),
+                ),
+                PopupMenuItem(
+                  value: "unsubscribe",
+                  child: Text("Unsubscribe to topic"),
+                ),
+                PopupMenuItem(
+                  value: "get_apns_token",
+                  child: Text("Get APNs token (Apple only)"),
+                ),
+              ];
+            },
+          ),
+        ],
       ),
       floatingActionButton: Builder(
         builder: (context) => FloatingActionButton(