+++ packages/firebase_ml_custom/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.4
+
+ - **FEAT**: bump compileSdkVersion to 29 (#3975).
+ - **FEAT**: bump `compileSdkVersion` to 29 in preparation for upcoming Play Store requirement.
+ - **CHORE**: publish packages.
+ - **CHORE**: publish packages.
+
 ## 0.0.3
 
  - **FEAT**: bump compileSdkVersion to 29 (#3975).